First Commit
So far I have made it so that most of the motors for the large
contraptions on the robot are up to date with the code in RobotMap. Bust
most importantly I have began work integrating the old swerve drive code
onto the new robot, this has been fairly straight forward with the
exception of now being able to use the Math.utils with java and not
needing the sun sqwauk version. I will finish converting it later on
tonight.