Made changes to integrate swerves
For simplicity sake I was forced to take out some of the smartdashboard
variables as they will not work until i fully integrate the code with
the OI and make the axis. So I commented almost all of the actual
driving out but if need be I can change that easily now. I also finished
integrating the swerve classes and changed everything to be compatible.